Bitwise Conversion Subroutines

Units of spatial and temporal dimensions, and MRI-specific spatial and temporal information.

xyzt2space(xyzt)

xyzt2time(xyzt)

space.time2xyzt(ss, tt)

dim2freq(di)

dim2phase(di)

dim2slice(di)

Arguments

xyzt
represents the units of pixdim[1..4] in the NIfTI header.

ss
is the character string of spatial units. Valid strings are: “Unknown”, “meter”, “mm” and “micron”.

tt
is the character string of temporal units. Valid strings are: “sec”, “msec”, “usec”, “Hz”, “ppm” and “rads”.

di
represents MRI slice ordering in the NIfTI header.

Value

For diminfo: the frequency, phase and slice dimensions encode which spatial dimension (1,2, or 3) corresponds to which acquisition dimension for MRI data. For xyzt_units: the codes are used to indicate the units of pixdim. Dimensions 1,2,3 are for x,y,z; dimension 4 is for time (t).

Details

The functions xyzt2space and xyzt2time can be used to mask off the undesired bits from the xyzt_units fields, leaving “pure” space and time codes.
http://nifti.nimh.nih.gov/nifti-1/documentation/nifti1fields/nifti1fields_pages/xyzt_units.html

The functions dim2freq, dim2phase, and dim2slice can be used to extract values from the dim_info byte.
http://nifti.nimh.nih.gov/nifti-1/documentation/nifti1fields/nifti1fields_pages/dim_info.html

References

Neuroimaging Informatics Technology Initiative (NIfTI)
http://nifti.nimh.nih.gov/

See also