Force an Object to Belong to the ANALYZE or NIfTI Class

Methods for function coerce in Package ‘methods’.

Arguments

object

is an object of class array or inherits from array.

Class

is the name of the class to which ‘object’ should be coerced; i.e., nifti.

from

is the object to be converted.

value

is the nifti class object to use as a template for various ANALYZE/NIfTI header information.

verbose

is a logical variable (default = FALSE) that allows text-based feedback during execution of the function.

Value

An object of class anlz or nifti.

Methods

from = "anlz", to = "nifti"

An object of class anlz is coerced into a NIfTI object.

from = "array", to = "anlz"

An object of class array is coerced into an ANALYZE object.

from = "array", to = "nifti"

An object of class array is coerced into a NIfTI object.

from = "list", to = "anlz"

All objects of class array in the list are coerced into ANALYZE objects. All other objects are left alone. The original list structure is retained.

from = "list", to = "nifti"

All objects of class array in the list are coerced into NIfTI objects. All other objects are left alone. The original list structure is retained.

See also

Author

Andrew Thornton zeripath@users.sourceforge.net,
Brandon Whitcher bwhitcher@gmail.com