Methods for Function `image'

Produce “lightbox” layout of images for nifti, anlz and afni objects.

# S4 method for nifti
image(
  x,
  z = 1,
  w = 1,
  col = gray(0:64/64),
  plane = c("axial", "coronal", "sagittal"),
  plot.type = c("multiple", "single"),
  zlim = NULL,
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  ...
)

# S4 method for anlz
image(
  x,
  z = 1,
  w = 1,
  col = gray(0:64/64),
  plane = c("axial", "coronal", "sagittal"),
  plot.type = c("multiple", "single"),
  zlim = NULL,
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  ...
)

# S4 method for afni
image(x, ...)

Arguments

x

is an object of class nifti or similar.

z

is the slice to be displayed (ignored when plot.type = "multiple").

w

is the time point to be displayed (4D arrays only).

col

is grayscale (by default).

plane

is the plane of acquisition to be displayed (choices are ‘axial’, ‘coronal’, ‘sagittal’).

plot.type

allows the choice between all slices being displayed, in a matrix (left-to-right, top-to-bottom), or a single slice.

zlim

is set to NULL by default and utilizes the internal image range.

xlab

is set to “” since all margins are set to zero.

ylab

is set to “” since all margins are set to zero.

axes

is set to FALSE since all margins are set to zero.

oma

is the size of the outer margins in the par function.

mar

is the number of lines of margin in the par function.

bg

is the background color in the par function.

...

other arguments to the image function may be provided here.

Details

Uses the S3 generic function image, with medical-image friendly settings, to display nifti, anlz and afni class objects in a “lightbox” layout.

Methods

x = "ANY"

Generic function: see image.

x = "nifti"

Produce images for x.

x = "anlz"

Produce images for x.

x = "afni"

Produce images for x.

See also

Author

Brandon Whitcher bwhitcher@gmail.com