Class "niftiAuditTrail"

An extension of the NIfTI class that adds an audit trail in XML format.

Objects from the Class

Objects can be created by calls of the form new("niftiAuditTrail", data, dim, dimnames, ...).

Methods

show

signature(object = "niftiAuditTrail"): prints out a summary of the imaging data.

References

NIfTI-1
http://nifti.nimh.nih.gov/

See also

Examples

showClass("niftiAuditTrail")

#> Class "niftiAuditTrail" [package "oro.nifti"]
#> 
#> Slots:
#>                                                                   
#> Name:           .Data          trail     extensions     sizeof_hdr
#> Class:          array            ANY           list        numeric
#>                                                                   
#> Name:       data_type        db_name        extents  session_error
#> Class:      character      character        numeric        numeric
#>                                                                   
#> Name:         regular       dim_info           dim_      intent_p1
#> Class:      character      character         vector        numeric
#>                                                                   
#> Name:       intent_p2      intent_p3    intent_code       datatype
#> Class:        numeric        numeric        numeric        numeric
#>                                                                   
#> Name:          bitpix    slice_start         pixdim     vox_offset
#> Class:        numeric        numeric         vector        numeric
#>                                                                   
#> Name:       scl_slope      scl_inter      slice_end     slice_code
#> Class:        numeric        numeric        numeric        numeric
#>                                                                   
#> Name:      xyzt_units        cal_max        cal_min slice_duration
#> Class:        numeric        numeric        numeric        numeric
#>                                                                   
#> Name:         toffset          glmax          glmin        descrip
#> Class:        numeric        numeric        numeric      character
#>                                                                   
#> Name:        aux_file     qform_code     sform_code      quatern_b
#> Class:      character        numeric        numeric        numeric
#>                                                                   
#> Name:       quatern_c      quatern_d      qoffset_x      qoffset_y
#> Class:        numeric        numeric        numeric        numeric
#>                                                                   
#> Name:       qoffset_z         srow_x         srow_y         srow_z
#> Class:        numeric         vector         vector         vector
#>                                                                   
#> Name:     intent_name          magic       extender     reoriented
#> Class:      character      character         vector        logical
#> 
#> Extends: 
#> Class "niftiExtension", directly
#> Class "nifti", by class "niftiExtension", distance 2
#> Class "array", by class "niftiExtension", distance 3
#> Class "structure", by class "niftiExtension", distance 4
#> Class "vector", by class "niftiExtension", distance 5, with explicit coerce