Methods for Function overlay

Methods for function overlay

overlay.nifti(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  plot.type = c("multiple", "single"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = FALSE,
  ...
)

overlay(x, y, ...)

# S4 method for nifti,missing
overlay(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  plot.type = c("multiple", "single"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = FALSE,
  ...
)

# S4 method for nifti,nifti
overlay(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  plot.type = c("multiple", "single"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = FALSE,
  ...
)

# S4 method for anlz,anlz
overlay(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  plot.type = c("multiple", "single"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = FALSE,
  ...
)

# S4 method for anlz,nifti
overlay(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  plot.type = c("multiple", "single"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = FALSE,
  ...
)

# S4 method for nifti,anlz
overlay(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  plot.type = c("multiple", "single"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = FALSE,
  ...
)

# S4 method for array,array
overlay(x, y, ...)

# S4 method for array,nifti
overlay(x, y, ...)

# S4 method for nifti,array
overlay(x, y, ...)

# S4 method for array,anlz
overlay(x, y, ...)

# S4 method for anlz,array
overlay(x, y, ...)

# S4 method for afni,afni
overlay(x, y, ...)

# S4 method for afni,array
overlay(x, y, ...)

Arguments

x, y
is an object of class nifti or similar.

z
is the slice to be displayed (ignored when plot.type = "multiple").

w
is the time point to be displayed (4D arrays only).

col.x
is grayscale (by default).

col.y
is hotmetal (by default).

zlim.x, zlim.y
are set to NULL (by default) and taken from the header information.

plane
is the plane of acquisition to be displayed (choices are ‘axial’, ‘coronal’, ‘sagittal’).

plot.type
allows the choice between all slices being displayed, in a matrix (left-to-right, top-to-bottom), or a single slice.

xlab
is set to “” since all margins are set to zero.

ylab
is set to “” since all margins are set to zero.

axes
is set to FALSE since all margins are set to zero.

oma
is the size of the outer margins in the par function.

mar
is the number of lines of margin in the par function.

bg
is the background color in the par function.

NA.x
Set any values of 0 in x to NA

NA.y
Set any values of 0 in y to NA

...
other arguments to the image function may be provided here.

Details

The image command is used multiple times to simultaneously visualize one of the three orthogonal planes in two multidimensional arrays, one on top of the other, for medical imaging data.

Methods

x = "nifti", y = "nifti"

Produce overlay of y on x.

x = "anlz", y = "anlz"

Produce overlay of y on x.

x = "afni", y = "afni"

Produce overlay of y on x.

See also