Methods for Function slice_overlay

Methods for function slice_overlay

slice_overlay.nifti(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = TRUE,
  useRaster = TRUE,
  ...
)

slice_overlay(x, y, ...)

# S4 method for class 'nifti,missing'
slice_overlay(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = TRUE,
  useRaster = TRUE,
  ...
)

# S4 method for class 'nifti,nifti'
slice_overlay(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = TRUE,
  useRaster = TRUE,
  ...
)

# S4 method for class 'anlz,anlz'
slice_overlay(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = TRUE,
  useRaster = TRUE,
  ...
)

# S4 method for class 'anlz,nifti'
slice_overlay(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = TRUE,
  useRaster = TRUE,
  ...
)

# S4 method for class 'nifti,anlz'
slice_overlay(
  x,
  y,
  z = 1,
  w = 1,
  col.x = gray(0:64/64),
  col.y = hotmetal(),
  zlim.x = NULL,
  zlim.y = NULL,
  plane = c("axial", "coronal", "sagittal"),
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  NA.x = FALSE,
  NA.y = TRUE,
  useRaster = TRUE,
  ...
)

# S4 method for class 'array,array'
slice_overlay(x, y, ...)

# S4 method for class 'array,nifti'
slice_overlay(x, y, ...)

# S4 method for class 'nifti,array'
slice_overlay(x, y, ...)

# S4 method for class 'array,anlz'
slice_overlay(x, y, ...)

# S4 method for class 'anlz,array'
slice_overlay(x, y, ...)

# S4 method for class 'afni,afni'
slice_overlay(x, y, ...)

# S4 method for class 'afni,array'
slice_overlay(x, y, ...)

Arguments

x, y

is an object of class nifti or similar.

z

is the slice to be displayed (ignored when plot.type = "multiple").

w

is the time point to be displayed (4D arrays only).

col.x

is grayscale (by default).

col.y

is hotmetal (by default).

zlim.x, zlim.y

are set to NULL (by default) and taken from the header information.

plane

is the plane of acquisition to be displayed (choices are ‘axial’, ‘coronal’, ‘sagittal’).

xlab

is set to “” since all margins are set to zero.

ylab

is set to “” since all margins are set to zero.

axes

is set to FALSE since all margins are set to zero.

oma

is the size of the outer margins in the par function.

mar

is the number of lines of margin in the par function.

bg

is the background color in the par function.

NA.x

Set any values of 0 in x to NA

NA.y

Set any values of 0 in y to NA

useRaster

if TRUE, a bitmap raster is used to plot the image instead of polygons. Passed to image

...

other arguments to the image function may be provided here.

Details

The image command is used multiple times to simultaneously visualize one of the three orthogonal planes in two multidimensional arrays, one on top of the other, for medical imaging data.

Methods

x = "nifti", y = "nifti"

Produce slice_overlay of y on x.

x = "anlz", y = "anlz"

Produce slice_overlay of y on x.

x = "afni", y = "afni"

Produce slice_overlay of y on x.

See also

Author

Brandon Whitcher bwhitcher@gmail.com