Methods for Function `orthographic' in Package `dcemriS4'

Produce orthographic display for nifti, anlz and afni objects.

orthographic.nifti(
  x,
  y = NULL,
  xyz = NULL,
  w = 1,
  col = gray(0:64/64),
  col.y = hotmetal(),
  zlim = NULL,
  zlim.y = NULL,
  crosshairs = TRUE,
  col.crosshairs = "red",
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  text = NULL,
  text.color = "white",
  text.cex = 2,
  ...
)

orthographic(x, ...)

# S4 method for nifti
orthographic(
  x,
  y = NULL,
  xyz = NULL,
  w = 1,
  col = gray(0:64/64),
  col.y = hotmetal(),
  zlim = NULL,
  zlim.y = NULL,
  crosshairs = TRUE,
  col.crosshairs = "red",
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  text = NULL,
  text.color = "white",
  text.cex = 2,
  ...
)

# S4 method for anlz
orthographic(
  x,
  y = NULL,
  xyz = NULL,
  w = 1,
  col = gray(0:64/64),
  col.y = hotmetal(),
  zlim = NULL,
  zlim.y = NULL,
  crosshairs = TRUE,
  col.crosshairs = "red",
  xlab = "",
  ylab = "",
  axes = FALSE,
  oma = rep(0, 4),
  mar = rep(0, 4),
  bg = "black",
  text = NULL,
  text.color = "white",
  text.cex = 2,
  ...
)

# S4 method for array
orthographic(x, ...)

# S4 method for afni
orthographic(x, ...)

Arguments

x
is an object of class nifti or similar.

y
is an object of class nifti or similar for the overlay.

xyz
is the coordinate for the center of the crosshairs.

w
is the time point to be displayed (4D arrays only).

col
is grayscale (by default).

col.y
is hotmetal (by default).

zlim
is the minimum and maximum ‘z’ values passed into image.

zlim.y
is the minimum and maximum ‘z’ values passed into image for the overlay.

crosshairs
is a logical value for the presence of crosshairs in all three orthogonal planes (default = TRUE).

col.crosshairs
is the color of the crosshairs (default = red).

xlab
is set to "" since all margins are set to zero.

ylab
is set to "" since all margins are set to zero.

axes
is set to FALSE since all margins are set to zero.

oma
is the size of the outer margins in the par function.

mar
is the number of lines of margin in the par function.

bg
is the background color in the par function.

text
allows the user to specify text to appear in the fourth (unused) pane.

text.color
is the color of the user-specified text (default = “white”.

text.cex
is the size of the user-specified text (default = 2).

...
other arguments to the image function may be provided here.

Methods

x = "afni"

Produce orthographic display for x.

x = "anlz"

Produce orthographic display for x.

x = "array"

Produce orthographic display for x.

x = "nifti"

Produce orthographic display for x.

See also